Revert "Renamed some functions, improved build system."
This reverts commit 0a2246a89962b0f21629971970a3532104433c8e.

+++ Minecraft/Engine/Aurealis/Graphics/Camera.h
@@ -48,12 +48,12 @@ namespace Aurealis
 			Timer passiveTimer;
 
 			float tiltAngle;
-			double bobY;
-			double bobElapsed;
+			float bobY;
+			float bobElapsed;
 			void update(bool isWalking);
 
 			float passiveElapsed;
-			double offAngleX, offAngleY;
+			float offAngleX, offAngleY;
 
 		//private:
